+++ Krowded-Mobile-App/Hello World App/src/main/AndroidManifest.xml
@@ -79,6 +79,10 @@
         <activity
             android:name=".ProfileActivity"
             android:label="@string/title_activity_profile"
+            android:theme="@style/AppTheme.NoActionBar" />
+        <activity
+            android:name=".CreateAccount"
+            android:label="@string/title_activity_create_account"
             android:theme="@style/AppTheme.NoActionBar"></activity>
     </application>
