+++ Krowded-Mobile-App/Hello World App/src/main/java/com/example/thebestteam/cs495capstonecomputing/MainActivity.java
@@ -17,7 +17,7 @@ import java.util.Objects;
 
 public class MainActivity extends AppCompatActivity {
     private boolean isFirstTime = true;
-    public User user;
+    public static User user;
 
     public User getUser() { return user; }
 
@@ -38,8 +38,6 @@ public class MainActivity extends AppCompatActivity {
             user = new User();
         }
 
-        Intent myIntent = new Intent(this, ProfileActivity.class);
-        startActivity(myIntent);
     }
 
 
@@ -65,8 +63,8 @@ public class MainActivity extends AppCompatActivity {
         EditText passBox = (EditText) findViewById(R.id.txtPassword);
         String password = passBox.getText().toString();
 
-        if (user.loggedin()) {
-            message = "You are already logged in, " + user.name() + ".";
+        if (user.getLoggedIn()) {
+            mSessage = "You are already logged in, " + user.getName() + ".";
         } else {
             if (Objects.equals(email, "krowded")) {
                 if (!Objects.equals(password, "123")) {
@@ -123,59 +121,8 @@ public class MainActivity extends AppCompatActivity {
 
 
     public void createAccount(View view) throws SQLException {
-        Context context = getApplicationContext();
-        int duration = Toast.LENGTH_SHORT;
-        CharSequence message;
-
-        EditText emailBox = (EditText) findViewById(R.id.txtEmail);
-        String email = emailBox.getText().toString();
-
-        EditText passBox = (EditText) findViewById(R.id.txtPassword);
-        String password = passBox.getText().toString();
-
-        if (Objects.equals(email, "krowded")) {
-            message = "An account with that email already exists.";
-        }
-        else {
-            message = "Account created. Welcome, " + email + "!";
-            user.logIn("TesterBoi",email,20,1,0);
-            Intent myIntent = new Intent(this, ProfileActivity.class);
-            startActivity(myIntent);
-        }
-
-
-        Toast toast = Toast.makeText(context,message,duration);
-        toast.show();
-
-        /*
-        // Send credentials to MariaDB. If the name is taken, toast "Try Again!", else create.
-        // Create the DB object
-        // Ordinarily you don't hardcode the password, of course
-        Connection conn = DriverManager.getConnection(
-                "jdbc:mysql://krowdeddb.cvnoof9d93qc.us-east-2.rds.amazonaws.com:3306/krowded", "krowded", "krowded4pp");
-
-        // Step 2: Allocate a 'Statement' object in the Connection
-        Statement stmt = conn.createStatement();
-
-        String query = "SELECT email, password FROM user WHERE email='test.me'";
-
-        ResultSet results = stmt.executeQuery(query);
-        CharSequence createResult;
-
-
-        if (results.isBeforeFirst() ) {
-            createResult = "Account already exists with that email.";
-            // createResult = "Account already exists with email '" + emailAddress + "'.";
-        }
-        else createResult = "I should create an account with that info.";
-
-        Context context = getApplicationContext();
-        int duration = Toast.LENGTH_LONG;
-
-        Toast toast = Toast.makeText(context,createResult,duration);
-        toast.show();
-
-    */
+        Intent myIntent = new Intent(this, CreateAccount.class);
+        startActivity(myIntent);
     }
 
 }
