+++ Krowded-Mobile-App/Hello World App/src/main/res/values/strings.xml
@@ -24,5 +24,6 @@
         <item>Restaurant</item>
     </string-array>
     <string name="title_activity_profile">My Profile</string>
+    <string name="title_activity_create_account">CreateAccount</string>
     <!-->>>>>>> 4c132646770886b0b9adf26c1908f8fc67d1b713-->
 </resources>
